2 examples - PHP upgrade + code quality
