+++ src/ValueObjects/AnotherClass.php
@@ -1,0 +1,8 @@
+<?php declare(strict_types=1);
+
+namespace App\ValueObjects;
+
+final class AnotherClass
+{
+
+}
